Remove method channel
