+++ packages/flutter/android/src/main/kotlin/io/sentry/flutter/SentryFlutterReplayRecorder.kt
@@ -3,7 +3,6 @@ package io.sentry.flutter
 import android.os.Handler
 import android.os.Looper
 import android.util.Log
-import io.flutter.plugin.common.MethodChannel
 import io.sentry.Sentry
 import io.sentry.protocol.SentryId
 import io.sentry.android.replay.Recorder
@@ -11,84 +10,76 @@ import io.sentry.android.replay.ReplayIntegration
 import io.sentry.android.replay.ScreenshotRecorderConfig
 
 internal class SentryFlutterReplayRecorder(
-  private val channel: MethodChannel,
+  private val callbacks: ReplayRecorderCallbacks,
   private val integration: ReplayIntegration,
 ) : Recorder {
+  private val main = Handler(Looper.getMainLooper())
+
   override fun start() {
-    Handler(Looper.getMainLooper()).post {
+    main.post {
       try {
         val replayId = integration.getReplayId().toString()
-        var replayIsBuffering = false
+        var buffering = false
         Sentry.configureScope { scope ->
-          // Buffering mode: we have a replay ID but it's not set on scope yet
-          replayIsBuffering = scope.replayId == SentryId.EMPTY_ID
+          buffering = scope.replayId == SentryId.EMPTY_ID
         }
-        channel.invokeMethod(
-          "ReplayRecorder.start",
-          mapOf(
-            "replayId" to replayId,
-            "replayIsBuffering" to replayIsBuffering,
-          ),
-        )
-      } catch (ignored: Exception) {
-        Log.w("Sentry", "Failed to start replay recorder", ignored)
+        callbacks.replayStarted(replayId, buffering)
+      } catch (t: Throwable) {
+        Log.w("Sentry", "Failed to start replay recorder (JNI)", t)
       }
     }
   }
 
   override fun resume() {
-    Handler(Looper.getMainLooper()).post {
+    main.post {
       try {
-        channel.invokeMethod("ReplayRecorder.resume", null)
-      } catch (ignored: Exception) {
-        Log.w("Sentry", "Failed to resume replay recorder", ignored)
+        callbacks.replayResumed()
+      } catch (t: Throwable) {
+        Log.w("Sentry", "Failed to resume replay recorder (JNI)", t)
       }
     }
   }
 
   override fun onConfigurationChanged(config: ScreenshotRecorderConfig) {
-    Handler(Looper.getMainLooper()).post {
+    main.post {
       try {
-        channel.invokeMethod(
-          "ReplayRecorder.onConfigurationChanged",
-          mapOf(
-            "width" to config.recordingWidth,
-            "height" to config.recordingHeight,
-            "frameRate" to config.frameRate,
-          ),
+        callbacks.replayConfigChanged(
+          config.recordingWidth,
+          config.recordingHeight,
+          config.frameRate,
         )
-      } catch (ignored: Exception) {
-        Log.w("Sentry", "Failed to propagate configuration change to Flutter", ignored)
+      } catch (t: Throwable) {
+        Log.w("Sentry", "Failed to propagate configuration (JNI)", t)
       }
     }
   }
 
   override fun reset() {
-    Handler(Looper.getMainLooper()).post {
+    main.post {
       try {
-        channel.invokeMethod("ReplayRecorder.reset", null)
-      } catch (ignored: Exception) {
-        Log.w("Sentry", "Failed to reset replay recorder", ignored)
+        callbacks.replayReset()
+      } catch (t: Throwable) {
+        Log.w("Sentry", "Failed to reset replay recorder (JNI)", t)
       }
     }
   }
 
   override fun pause() {
-    Handler(Looper.getMainLooper()).post {
+    main.post {
       try {
-        channel.invokeMethod("ReplayRecorder.pause", null)
-      } catch (ignored: Exception) {
-        Log.w("Sentry", "Failed to pause replay recorder", ignored)
+        callbacks.replayPaused()
+      } catch (t: Throwable) {
+        Log.w("Sentry", "Failed to pause replay recorder (JNI)", t)
       }
     }
   }
 
   override fun stop() {
-    Handler(Looper.getMainLooper()).post {
+    main.post {
       try {
-        channel.invokeMethod("ReplayRecorder.stop", null)
-      } catch (ignored: Exception) {
-        Log.w("Sentry", "Failed to stop replay recorder", ignored)
+        callbacks.replayStopped()
+      } catch (t: Throwable) {
+        Log.w("Sentry", "Failed to stop replay recorder (JNI)", t)
       }
     }
   }
@@ -97,3 +88,5 @@ internal class SentryFlutterReplayRecorder(
     stop()
   }
 }
+
+
